pyproject.toml
[tool.poetry]
name = "digitalocean"
version = "0.0.1"
description = "Ansible DigitalOcean Collection"
authors = ["Mark Mercado <mmercado@digitalocean.com>", "Ari Kalfus <akalfus@digitalocean.com>"]
license = "GPL-3.0"
readme = "README.md"

[tool.poetry.dependencies]
python = "~3.9"
ansible-core = "^2.15.3"
azure-core = "^1.29.3"
pydo = "^0.1.4"
boto3 = "^1.28.35"
jmespath = "^1.0.1"

[tool.poetry.group.dev.dependencies]
ansible-lint = { version = "^6.16.2", markers = "platform_system != 'Windows'" }
black = "^23.3.0"
pre-commit = "^3.3.2"
pylint = "^2.17.4"
pytest = "^7.3.1"
antsibull-changelog = "^0.23.0"
antsibull-docs = "^2.3.1"
autopep8 = "^2.0.4"

[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
